Add files via upload
Background subtraction is a fundamental technique in computer vision for
distinguishing moving foreground objects from the static background in video
sequences. It is widely employed in surveillance systems, motion tracking, and
activity recognition. The aim is to generate a dynamic understanding of the
scene by isolating regions of interest (foreground) while maintaining an
accurate background model.
This project utilizes OpenCV's background subtraction methods to process a
video, generating separate videos for the moving foreground and the static
background.